requirements.txt
flask==3.0.0
yt-dlp==2024.1.7
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
